Update project Go toolchain version to 1.22.9
Support for specifying the Go toolchain version compatibility of a module to patch version granularity via the `go`
directive of the go.mod module metadata file was introduced in Go 1.21.0. The `go` directives of the project's modules
were changed to specify the patch version when the project's Go version was bumped to 1.22.3. However, support for this
version format was not added to the `go fix` command until Go 1.22.7.

This caused the `go fix` command (which is invoked by the "check-outdated" job of the project's "Check Go" workflow) to
fail when a version of Go between 1.21.0 and 1.22.6 was used:

```
invalid -go=go1.22.3
exit status 2
task: Failed to run task "go:fix": exit status 1
```

Previously, this did not affect the CI system because the workflows only specified the Go version to be installed by the
"actions/setup-go" action to the minor version (1.22), and the action happened to use a version of Go >1.22.6. However,
the "actions/setup-go" action now installs the exact version of Go specified the `go` directive, which means the runs of
the "Check Go" workflow would fail with the above error if the value of the `go` directive is >=1.21.0 and <1.22.7.

The chosen solution is to bump the value of the `go` directive to a version with the `go fix` command bug fixed.
Although this could be achieved by a bump to 1.22.7, it makes sense to bump Go all the way to the latest version in the
1.22.x series (because apparently the production release build system is not compatible with Go 1.23.x).