Update redirecting URLs in reference links
The templates and documentation contain reference links to provide additional information to the users and maintainers.

The targets of some of these links have moved since the time they were added. Although the user could still reach the
intended content via a redirect, it is best not to rely on redirects continuing to work indefinitely. So the URLs are
hereby updated to point directly to the target content.