Standardize formatting of task parameter documentation
Some tasks accept input via environment variables. These parameter variables are documented in the task description.

A standard format has been established for that documentation. Previously, these descriptions did not follow the
standardized format.

Typically, the documentation is done in the task description, to make the information easily accessible to contributors
via the `task list` output. However, a description was intentionally omitted for these tasks. The reason is these tasks
are for internal use by other tasks, only serving to avoid code duplication. So the tasks should not be listed in the
`task list` output. Task actually has a feature for marking such tasks as internal, but unfortunately that only works
for the standard way of calling a task from another task, which can not be used in this case where the output of the
task must be captured. For this reason, the task documentation is intentionally done via a comment instead of the
description as would be done for a contributor facing task.