Use project version of ajv-cli in npm:validate task
The "ajv-cli" command line tool is used for validating data files against their JSON schema. The tool is used to
validate the project's npm package.json configuration file.

In general, it is preferable (and for some schemas even mandatory) to use the latest version of ajv-cli. However,
support for the "Draft-04" schema specification was dropped in ajv-cli version 4.0.0. So when working with JSON schemas
that specify that draft, it is necessary to use ajv-cli 3.3.0, the last compatible version.

Previously, the package.json schema specified the "Draft-04" specification. For this reason, the `npm:validate` task was
configured to use ajv-cli@3.3.0.

The package.json schema has now been updated to use the "Draft-07" schema specification. So the code for using
ajv-cli@3.3.0 is removed from the task, and it will now instead use the standard project level version of ajv-cli.