Only install Python package dependencies from relevant group
The "Poetry" tool is used to manage the project's Python package dependencies.

Dependencies might be classified into distinct categories. The most basic classification would be:

- Application dependencies: used by the project's applications
- Development dependencies: tools used in the development and maintenance of the project, but not by the application

By default, Poetry installs all non-optional dependencies. This can be inefficient in a case where a specific operation
is being performed, since a given operation might only require the dependencies from one category and so the
installation of dependencies from the other is pointless for that operation.

For this reason, Poetry allows the user to organize dependencies into arbitrary "groups", and to specify which groups
should be installed.

The Python package installation task is hereby updated to allow dependency groups to be specified, and the calls to that
task updated to specify the groups they require.