Wrap signing certificate password in single quotes to prevent corruption
High quality passwords may contain characters with special treatment by the shell (e.g., $).

The previous use of double quotes in the shell command for importing the macOS signing certificate to the runner's
keychain resulted in the password being corrupted, and thus a spurious failure in this case:

```
security: SecKeychainItemImport: The user name or passphrase you entered is not correct.
```