Skip publishing releases to AWS if credentials not configured
The "Release" and "Publish Nightly Build" workflows upload the generated build files to the AWS S3 bucket used by
Arduino's downloads server.

The necessary credentials are configured in Arduino's repository. However, these workflows might be used in other
contexts:

- by contributors validating proposed changes to the release infrastructure in their fork
- by hard forks of the project

In either case (especially the former), the fork owner is unlikely to be willing/able to set up the AWS infrastructure
that would be needed to use this capability of the workflow.

Since these workflows also publish the builds to GitHub, the AWS upload is not essential to either 3rd party use case.

The workflow code is hereby configured to skip the AWS upload steps if the necessary credentials have not been
configured in the repository. The existence of the `AWS_ROLE_TO_ASSUME` repository secret is used as the indicator of
whether the credentials are configured. This will allow runs of the workflow in forks without the need to remove the AWS
upload steps.